+++ evm/test/SignedSafeMath.test.ts
@@ -29,7 +29,7 @@ describe('SignedSafeMath', () => {
   const INT256_MIN = ethers.utils.bigNumberify(
     '-57896044618658097711785492504343953926634992332820282019728792003956564819968',
   )
-  const deployment = h.useSnapshot(provider, async () => {
+  const deployment = providers.useSnapshot(provider, async () => {
     adder = await concreteSignedSafeMathFactory.connect(defaultAccount).deploy()
   })
